requirements.txt
gradio
transformers
torch
Pillow
pandas
numpy